+++ src/muc_light/mod_muc_light_codec_modern.erl
@@ -58,12 +58,11 @@ encode({#msg{} = Msg, AffUsers}, Sender, {RoomU, RoomS} = RoomUS, HandleFun) ->
              {<<"from">>, RoomBin}
             ],
     MsgForArch = #xmlel{ name = <<"message">>, attrs = Attrs, children = Msg#msg.children },
-    EventData = [{from_nick, FromNick},
-                 {from_jid, Sender},
-                 {room_jid, jid:make_noprep({RoomU, RoomS, <<>>})},
-                 {affiliation, Aff},
-                 {role, mod_muc_light_utils:light_aff_to_muc_role(Aff)}
-                ],
+    EventData = #{from_nick =>FromNick,
+                  from_jid => Sender,
+                  room_jid => jid:make_noprep({RoomU, RoomS, <<>>}),
+                  affiliation => Aff,
+                  role => mod_muc_light_utils:light_aff_to_muc_role(Aff)},
     #xmlel{ children = Children }
                      = mongoose_hooks:filter_room_packet(RoomS, MsgForArch, EventData),
     lists:foreach(
@@ -511,9 +510,10 @@ b2what(<<"room">>) -> room.
 what2b(user) -> <<"user">>;
 what2b(room) -> <<"room">>.
 
+-spec room_event(jid:jid()) -> mod_muc:room_event_data().
 room_event(RoomJID) ->
-    [{from_nick, <<>>},
-     {from_jid, RoomJID},
-     {room_jid, RoomJID},
-     {role, owner},
-     {affiliation, owner}].
+    #{from_nick => <<>>,
+      from_jid => RoomJID,
+      room_jid => RoomJID,
+      role => owner,
+      affiliation => owner}.